updated auth service and functionality and files hierarchy
